CORE-760 Cleaned up the LMDBOps code and comment explaining key encoding strategy.
